refactor: move `INITIAL_EXPECTED_NAMES` to the crate root
The constant is no longer used in `backend.rs` itself. Its three users
are the Crankshaft name generators in `backend/docker.rs` and
`backend/tes.rs` and the evaluator's own generator in `eval/v1.rs`, so it
now sits beside `ONE_GIBIBYTE` at the crate root, where a private const
is visible to every descendant module without the `pub(crate)` this
branch had added.

Both status migrations now state why they decline sqlx's wrapping
transaction rather than pointing at each other. Testing the claim against
a seeded database showed that deferring foreign key enforcement is not an
alternative to disabling it: the commit fails with `FOREIGN KEY
constraint failed`, because renaming the replacement table into place
does not clear the violations that dropping the original recorded.

The changelog entries for this branch are shorter and cite the pull
request.
