+++ crates/wdl-engine/src/backend/docker.rs
@@ -39,12 +39,12 @@ use crate::CLEANUP_TASK_NAME_PREFIX;
 use crate::CancellationContext;
 use crate::EvaluationPath;
 use crate::Events;
+use crate::INITIAL_EXPECTED_NAMES;
 use crate::ONE_GIBIBYTE;
 use crate::Object;
 use crate::PrimitiveValue;
 use crate::TaskInputs;
 use crate::backend::ExecuteTaskRequest;
-use crate::backend::INITIAL_EXPECTED_NAMES;
 use crate::backend::manager::TaskManager;
 use crate::config::Config;
 use crate::config::TaskResourceLimitBehavior;
