+++ crates/wdl-engine/src/backend/tes.rs
@@ -41,11 +41,11 @@ use crate::EngineEvent;
 use crate::EvaluationPath;
 use crate::EvaluationPathKind;
 use crate::Events;
+use crate::INITIAL_EXPECTED_NAMES;
 use crate::ONE_GIBIBYTE;
 use crate::Object;
 use crate::PrimitiveValue;
 use crate::TaskInputs;
-use crate::backend::INITIAL_EXPECTED_NAMES;
 use crate::backend::STDERR_FILE_NAME;
 use crate::backend::STDOUT_FILE_NAME;
 use crate::backend::WORK_DIR_NAME;
