+++ crates/wdl-engine/src/eval/v1.rs
@@ -27,8 +27,8 @@ use wdl_analysis::types::EnumChoiceCacheKey;
 use super::CancellationContext;
 use super::Events;
 use crate::EngineEvent;
+use crate::INITIAL_EXPECTED_NAMES;
 use crate::Value;
-use crate::backend::INITIAL_EXPECTED_NAMES;
 use crate::backend::TaskExecutionBackend;
 use crate::cache::CallCache;
 use crate::cache::CallCacheExclusions;
